backend/requirements.txt
Django==4.2.14
djangorestframework
django-cors-headers
psycopg2-binary
python-decouple
channels
daphne
channels-redis
asyncio
requests
pyotp
djangorestframework-simplejwt
pillow
django-silk
faker
qrcode